chore: update on web tips
