+++ packages/kit/src/provider/Container/ErrorToastContainer/ErrorToastContainer.tsx
@@ -1,22 +1,18 @@
 import { useEffect } from 'react';
 
-import { useIntl } from 'react-intl';
-
-import { Button, Toast, useClipboard } from '@onekeyhq/components';
+import { Toast } from '@onekeyhq/components';
 import type { IAppEventBusPayload } from '@onekeyhq/shared/src/eventBus/appEventBus';
 import {
   EAppEventBusNames,
   appEventBus,
 } from '@onekeyhq/shared/src/eventBus/appEventBus';
-import { ETranslations } from '@onekeyhq/shared/src/locale';
-import { isRequestIdMessage } from '@onekeyhq/shared/src/request/utils';
+
+import { getErrorAction } from './ErrorToasts';
 
 const ERROR_CODE = [403];
 const isFilterErrorCode = (code?: number) => code && ERROR_CODE.includes(code);
 
 export function ErrorToastContainer() {
-  const intl = useIntl();
-  const { copyText } = useClipboard();
   useEffect(() => {
     const fn = (p: IAppEventBusPayload[EAppEventBusNames.ShowToast]) => {
       const message = p.message;
@@ -24,18 +20,8 @@ export function ErrorToastContainer() {
         ? String(p.errorCode)
         : undefined;
       toastId = toastId || (p.title ? p.title : message);
-      const actions = isRequestIdMessage(message) ? (
-        <Button
-          size="small"
-          onPress={() => {
-            if (message) {
-              copyText(message);
-            }
-          }}
-        >
-          {intl.formatMessage({ id: ETranslations.global_copy })}
-        </Button>
-      ) : undefined;
+      const actions = getErrorAction(p.errorCode, message ?? '');
+
       Toast[p.method]({
         ...p,
         toastId,
@@ -46,7 +32,7 @@ export function ErrorToastContainer() {
     return () => {
       appEventBus.off(EAppEventBusNames.ShowToast, fn);
     };
-  }, [copyText, intl]);
+  }, []);
 
   return null;
 }
